% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ismb_dat2019}
\alias{ismb_dat2019}
\title{Tweets from ISMBECCB, 2019 conference scraped from twitter on 25th July evening after
the last keynote}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3246 rows and 3 columns.}
\usage{
ismb_dat2019
}
\description{
Tweets from ISMBECCB, 2019 conference scraped from twitter on 25th July evening after
the last keynote
}
\keyword{datasets}
